Diffusive model

Function to evaluate the influence area of a stressor using a passive diffusive model

diffusive(dat, field, threshold, globalmaximum, decay, distance, increment)

Arguments

dat

sf object with points characterizing stressor intensity

field

character, name of field containing stresseor intensity

threshold

numeric, minimum threshold in percent of the global maximum at which to stop the diffusive model

globalmaximum

global maximum value of the stressor in the study area

decay

percent by which the value of the stressor is reduced when diffusing to the adjacent cells

distance

numeric, distance beyond which the stressor is considered to have no influence. Units are in the units of the spatial object provided.

increment

numeric, distance between successive steps in the model. Units are in the units of the spatial object provided.

Details